Fixing the following issues - turning smart quotes, and long dash replacement OFF
1. replacing smart quotes by straight quotes 
2. replacing long dash back to two dashes -- (this was problematic in
with xml comments: '<!--'  3. adhering to sequences as specified in the
xsd  4. Removing spaces from wrapped text. id, link and href fields
ended up with white space in them. The formatting does not look as good
now, but at least the xml is valid now.  5. xmlSchemaLocation ended up
with two href's: fixed this back to one.  6. Checked the HTTP response
codes in the examples.